fix(): rename component
