+++ src/components/LineChartComponent.tsx
